rename component classifications
