+++ src/README.md
@@ -1,0 +1,6 @@
+# Components
+
+- **Independent components** can import NPM packages and CSS
+- **Basic components** can additionally import independent components
+- **Complex components** can additionally import basic components, anc connect to store
+- **Screen components** can additionally be set to `<Route>`
