+++ src/complexes/DefaultLayout.test.tsx
@@ -2,8 +2,8 @@ import { mount, ReactWrapper } from 'enzyme';
 import firebase from 'firebase';
 import React from 'react';
 import { BrowserRouter } from 'react-router-dom';
-import Footer from '../components/Footer';
-import Header from '../components/Header';
+import Footer from '../basics/Footer';
+import Header from '../basics/Header';
 import { IUserProfile } from '../reducers/currentUser';
 import { DefaultLayout } from './DefaultLayout';
