+++ src/independents/LoadingView.tsx
