+++ src/screens/HomePage.tsx
@@ -3,11 +3,11 @@ import { connect } from 'react-redux';
 import { Link } from 'react-router-dom';
 import styled from 'styled-components';
 import logoImageSrc from '../assets/logo-512.png';
-import Container from '../basics/Container';
-import LoadingView from '../basics/LoadingView';
-import Footer from '../components/Footer';
-import Header from '../components/Header';
-import { getDefaultHeaderMenu } from '../containers/DefaultLayout';
+import Footer from '../basics/Footer';
+import Header from '../basics/Header';
+import { getDefaultHeaderMenu } from '../complexes/DefaultLayout';
+import Container from '../independents/Container';
+import LoadingView from '../independents/LoadingView';
 import { store } from '../misc';
 import { connectActiveRooms } from '../models/rooms';
 import path, { RoomLink } from '../path';
