+++ src/screens/LogoutPage.tsx
@@ -1,5 +1,5 @@
 import React, { useState } from 'react';
-import DefaultLayout from '../containers/DefaultLayout';
+import DefaultLayout from '../complexes/DefaultLayout';
 import * as users from '../models/users';
 import { HomeLink } from '../path';
