+++ src/screens/RoomSettingsPage.tsx
@@ -2,9 +2,9 @@ import React, { useState } from 'react';
 import { connect } from 'react-redux';
 import { RouteComponentProps } from 'react-router-dom';
 import styled from 'styled-components';
-import Footer from '../components/Footer';
-import RoomHeader from '../components/RoomHeader';
-import { MainContainer } from '../containers/DefaultLayout';
+import Footer from '../basics/Footer';
+import RoomHeader from '../basics/RoomHeader';
+import { MainContainer } from '../complexes/DefaultLayout';
 import { appHistory } from '../misc';
 import { RoomLink } from '../path';
 import { Dispatch, IState } from '../reducers';
