+++ src/screens/RoomTextbookPage.tsx
@@ -2,12 +2,12 @@ import React, { useState } from 'react';
 import { connect } from 'react-redux';
 import { RouteComponentProps } from 'react-router-dom';
 import styled from 'styled-components';
-import Container from '../basics/Container';
-import LoadingView from '../basics/LoadingView';
-import { headerHeight } from '../components/Header';
-import RoomHeader from '../components/RoomHeader';
-import TextbookContent from '../components/TextbookContent';
-import DefaultLayout from '../containers/DefaultLayout';
+import { headerHeight } from '../basics/Header';
+import RoomHeader from '../basics/RoomHeader';
+import TextbookContent from '../basics/TextbookContent';
+import DefaultLayout from '../complexes/DefaultLayout';
+import Container from '../independents/Container';
+import LoadingView from '../independents/LoadingView';
 import { store } from '../misc';
 import { connectActiveRooms } from '../models/rooms';
 import { IState } from '../reducers';
