+++ src/screens/RoomWritePage.tsx
@@ -3,9 +3,9 @@ import React, { createRef } from 'react';
 import { connect } from 'react-redux';
 import { RouteComponentProps } from 'react-router-dom';
 import styled, { css } from 'styled-components';
-import { headerHeight } from '../components/Header';
-import RoomHeader from '../components/RoomHeader';
-import TextbookContent from '../components/TextbookContent';
+import { headerHeight } from '../basics/Header';
+import RoomHeader from '../basics/RoomHeader';
+import TextbookContent from '../basics/TextbookContent';
 import syncScroll from '../functions/syncScroll';
 import { debounce } from '../misc';
 import { Dispatch, IState } from '../reducers';
