+++ src/screens/SettingsPage.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { connect } from 'react-redux';
 import { Link } from 'react-router-dom';
-import DefaultLayout from '../containers/DefaultLayout';
+import DefaultLayout from '../complexes/DefaultLayout';
 import * as users from '../models/users';
 import path from '../path';
 import { Dispatch, IState } from '../reducers';
